requirements.txt
tensorly
pytorch